requirements.txt
requests==2.31.0
pathlib==1.0.1
python-dotenv==1.0.0